converters: Add debug logging
If a problem within the converter happens, it can be difficult
to understand what is the origin of the problem. In order to make
it easier to find out the actual issue, we add debug messages which
can be activated by setting the correct global logger level.

This also fixes https://github.com/mutwo-org/mutwo.midi/issues/3.