converters/EventToMidiFile: Fix crash if extractor returns None
If an extraction function (e.g. event->pitch_list, event->volume,
event->midi_control_tuple) returns 'None' (maybe because the events
attribute is set to 'None') the converter used to crash before this
patch.

Now it doesn't crash anymore, but a warning is raised that the given
event is interpreted as a rest.