converters/frontends: Cosmetics (improve code style)
Before this patch the code of the frontend converter was
extremely verbose, because no abbreviations were used in internal code.
This isn't necessary, the context is already clear within a code and
the extremly long variable names rather obscured the meaning instead
of helping it. So this rework makes the file much smaller (~100 lines
less code). It doesn't change any functionality, it only helps
maintaining the source code.